Changes
-sending agents' species number from rt to unity via agentIDs file.
- computing the hamilton satisfiability  and adding it to the input
array

+++ Assets/Scripts/NeuroEvolution.cs
@@ -34,7 +34,7 @@ public class NeuroEvolution : MonoBehaviour {
     void writeNNInput()
     {
         string lines = "";
-        float[] inputArray = GameObject.Find("Agent").GetComponent<Agent>().inputArray;
+        double[] inputArray = GameObject.Find("Agent").GetComponent<Agent>().inputArray;
         foreach (float input in inputArray)
         {
             lines += input + ",";
